+++ chebyshev_pseudospectral/src/ChebyshevDifferentiate.hpp
@@ -33,8 +33,8 @@
 
 void SpectralDifferentiate(Eigen::Ref<RowVectorXd> data,
                            Eigen::Ref<RowVectorXd> rslt,
-                           double span, unsigned n);
+                           double span, size_t n);
 
 void BoundaryConditions(Eigen::Ref<RowVectorXd> data,
                         Eigen::Ref<RowVectorXd> rslt,
-                        unsigned n);
+                        size_t n);
